Refactor StickyProductHeader
